==> analysis/1_GenerateMapAndParams.Rmd <==
---
title: "Generate map and calculate parameter values"
author: "E Schlatter"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
setwd("C:/Users/eschlatter/Dropbox/DispersalSGS/analysis")
library(tidyverse)
library(sf)
library(sp)
library(stars)
library(raster)
library(geosphere)
library(lwgeom)
```

### Import data: D'Aloia et al SGS sampling sites
```{r, eval=FALSE}
sgs_sites <- read.csv('../data/SGS_sites.csv')

sgs_points <- SpatialPoints(sgs_sites[,c('Coord_UTM_zone_16N.x','Coord_UTM_zone_16N.y')])
crs.geo <- CRS("+init=epsg:32616")  # https://spatialreference.org/ref/epsg/?search=utm+16N
proj4string(sgs_points) <- crs.geo  # define projection system of our data
sgs_points_df <- SpatialPointsDataFrame(sgs_points,sgs_sites)
sgs_points_sf <- st_as_sf(sgs_points_df)

spgeo <- spTransform(sgs_points, CRS("+proj=longlat +datum=WGS84"))
sites_lonlat <- coordinates(spgeo)
sgs_points_df <- cbind(sgs_points_df,sites_lonlat)
save(sgs_points,sgs_points_df,sgs_points_sf,sites_lonlat,file='../data/SGS_sites.RData')
```

### Import data: Global Distribution of Coral Reefs/Millennium Coral Reef Mapping Project
https://data.unep-wcmc.org/datasets/1
```{r, eval=FALSE}
load('../data/SGS_sites.RData')

# points with expert-verified habitat info, etc
# not much useful in our region
mmp_shp <- read_sf("../data/14_001_WCMC008_CoralReefs2021_v4_1/01_Data/WCMC008_CoralReef2021_Pt_v4_1.shp")  # points

ggplot(mmp_shp)+
  geom_sf(aes(color=LOC_DEF))+
  geom_sf(data=sgs_points_sf,shape='*',size=5)+
  theme_minimal()+
  xlim(c(-88.5,-87))+
  ylim(c(16,19))

# this is the one we want. Polygons/multipolygons classified as coral. Subclassifications don't seem very helpful.
mmpy_shp <- read_sf("../data/14_001_WCMC008_CoralReefs2021_v4_1/01_Data/WCMC008_CoralReef2021_Py_v4_1.shp")

# pull out the general area we want and save it
sf_use_s2(FALSE) # otherwise, trying to crop gives an error due to spherical geometry
bze <- st_crop(mmpy_shp,xmin=-88.5,xmax=-87,ymin=16,ymax=19)
rm(mmpy_shp) #it's big

save(bze,file='../data/mmp_bze.RData')
```
```{r}
library(marmap)
load('../data/SGS_sites.RData')

bathy_bze <- getNOAA.bathy(lon1 = -89, lon2 = -87, lat1 = 16, lat2 = 19, resolution = 0.25)
bathy_xyz <- as.xyz(bathy_bze)

shallows <- c("skyblue","skyblue1","skyblue2","skyblue3")
meds <- c("yellow1","yellow2","yellow3")
deeps <- c("dodgerblue","dodgerblue1","dodgerblue2","dodgerblue3")


blues <- c("lightsteelblue1", "lightsteelblue2", "lightsteelblue3", "lightsteelblue4")

plot(bathy_bze, ylim=c(16.5,18), image = TRUE, land = TRUE, lwd = 0.1, n=10,
     bpal = list(c(0, max(bathy_bze), "grey"),
                 c(0,min(bathy_bze),blues),
                 c(-30,-70,meds)))
plot(bathy_bze, deep = 0, shallow = 0, step = 0, lwd = 0.4, add = TRUE)
points(sites_lonlat[,1],sites_lonlat[,2],pch=21,col="black",bg="red",cex=0.8)
scaleBathy(bathy_bze, deg = 0.2, x = "bottomleft", inset = 5)


# plot(bathy_bze, image = TRUE, land = TRUE, lwd = 0.1, n=10,
# bpal = list(c(0, max(bathy_bze), "grey"),
#             c(0,-45,shallows),
#             c(-45,-90,meds),
#             c(-90,min(bathy_bze),deeps)))
# plot(bathy_bze, deep = 0, shallow = 0, step = 0,
# lwd = 0.4, add = TRUE)
# scaleBathy(bathy_bze, deg = 0.2, x = "bottomleft", inset = 5)

#plot(bathy_bze, image = TRUE)
#scaleBathy(bathy_bze, deg = 2, x = "bottomleft", inset = 5)
```

### Determine geographic extent for simulation
I'm using 16.4km (i.e., max goby dispersal distance) outside of the SGS sampling area in each direction. If we simulate in that area (and let individuals die if they leave it), we shouldn't have to worry too much about edge effects.
```{r, eval=FALSE}
load('../data/mmp_bze.RData')
load('../data/SGS_sites.RData')

#sites that are the farthest north (B1), south (B24), east (L4) and west (B25)
borders=sites_lonlat[c(1,39,33,40),]

# points 16.4km N,S,E,W of those (respectively)
bord_n <- destPoint(borders[1,],0,16.4*1000)
bord_s <- destPoint(borders[2,],180,16.4*1000)
bord_e <- destPoint(borders[3,],90,16.4*1000)
bord_w <- destPoint(borders[4,],270,16.4*1000)

# get the max lat and lon of those points. These will be the boundaries of our map.
bords <- rbind(bord_n,bord_s,bord_w,bord_e)
min_lat <- min(bords[,2])
max_lat <- max(bords[,2])
min_lon <- min(bords[,1])
max_lon <- max(bords[,1])

# generate a reef object cropped to that size
reef_sim <- st_crop(bze,xmin=min_lon,xmax=max_lon,ymin=min_lat,ymax=max_lat,crop=TRUE)

## get the size of the map, in meters
# the x-direction isn't exact, because we're on a sphere. The distance between the top two corners is about 1km less than the distance between the bottom two corners. I'm not sure how this gets resolved.
dist_x1 <- distGeo(c(min_lon,max_lat),c(max_lon,max_lat))
dist_x2 <- distGeo(c(min_lon,min_lat),c(max_lon,min_lat))
dist_x <- (dist_x1+dist_x2)/2/1000
    
# not a problem in the y-direction, because distance N-S is the same, no matter which longitude you're on
dist_y <- distGeo(c(min_lon,min_lat),c(min_lon,max_lat))/1000

save(reef_sim,min_lon,max_lon,min_lat,max_lat,dist_x,dist_y,file='../data/mmp_crop.RData')
```

### TNC Caribbean dataset:
https://hub.arcgis.com/datasets/TNC::coral-reefs-2017/explore
(Note: site is retiring; data also available at https://map.unbiodiversitylab.org/earth?basemap=grayscale&coordinates=17.0655843,-87.7560641,9&layers=global-distribution-of-warm-water-coral-reefs_100)
I'm going to use this dataset instead, at least for now, because the lagoon and interior atolls are cropped

```{r,eval=FALSE}
load('../data/mmp_crop.RData')

TNC <- read_sf('../data/TNC_CoralReefs.geojson')
TNC_sub <- filter(TNC,FID %in% c(24,25,26,29,31,59,64))  # include 59 (southern lagoon) for now, to make cropping possible, then remove it afterward
sf_use_s2(FALSE)
TNC_sim <- st_crop(TNC_sub,xmin=min_lon,xmax=max_lon,ymin=min_lat,ymax=max_lat,crop=TRUE)
TNC_sim <- filter(TNC_sim,FID %in% c(24,25,26,29,31))
save(TNC_sim,file='../data/TNC_sim.RData')
```

```{r,eval=FALSE}
load('../data/TNC_sim.RData')
load('../data/SGS_sites.RData')

## Get the barrier reef
barrier <- TNC_sim[c(1,2,3,4),]$geometry

## Get the atolls, cut out the holes, put them back into a multipolygon with CRS
atolls=TNC_sim[5,]$geometry

# glovers
glovers <- st_polygon(atolls[[1]][[1]])
glovers <- st_multipolygon(list(glovers)) %>% st_sfc(crs=4326)

# lighthouse
atoll2 <- st_polygon(atolls[[1]][[2]]) #outside
atoll3 <- st_polygon(atolls[[1]][[3]]) #inside
lighthouse <- st_difference(atoll2,atoll3)
lighthouse <- st_multipolygon(list(lighthouse)) %>% st_sfc(crs=4326)

# turneffe
atoll5 <- st_polygon(atolls[[1]][[5]]) # outside
atoll6 <- st_polygon(atolls[[1]][[6]]) # inside
turneffe <- st_difference(atoll5,atoll6)
turneffe <- st_multipolygon(list(turneffe)) %>% st_sfc(crs=4326)

b_color='#5e3c99' #purple
g_color='#e66101' #orange
l_color='#1b7837' #green

g_color= '#fc8d59' #orange (G)
l_color='#ffffbf' #yellow (L)
b_color='#91bfdb' #blue (B)

ggplot(barrier)+
  geom_sf(color='black',fill='black')+
  geom_sf(data=turneffe,color='black',fill='black')+
  geom_sf(data=glovers,color='black',fill='black')+
  geom_sf(data=lighthouse,color='black',fill='black')+
  geom_sf(data=sgs_points_sf,color='black',fill='white',shape=21,size=3)+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        axis.text=element_blank(),axis.ticks=element_blank(),
        panel.background = element_rect(fill = "#f0f0f0"))+
  ylim(16.42,18.16) +
  xlim(-88.2,-87.4)+
  geom_text(aes(x=-87.65,y=18.1),label="Barrier")+
  geom_text(aes(x=-87.7,y=17.7),label="Turneffe")+
  geom_text(aes(x=-87.8,y=16.65),label="Glover's")+
  geom_text(aes(x=-87.65,y=17.05),label="Lighthouse")+
  xlab(NULL)+ylab(NULL)

ggplot(sgs_points_sf,size=3, shape=21, aes(color=Region))+
  geom_sf()+
  scale_color_manual(values=c(b_color,b_color,b_color,g_color,l_color,b_color))+
  geom_sf(data=barrier,color='black',fill='black')

library(ggspatial)

######## broad scale colors map (for talk)

ggplot(data=barrier)+
  geom_sf(color='black',fill='black')+
  geom_sf(data=turneffe,color='black',fill='black')+
  geom_sf(data=glovers,color='black',fill='black')+
  geom_sf(data=lighthouse,color='black',fill='black')+
  geom_sf(data=sgs_points_sf,size=3, shape=21, aes(fill=Region))+
  scale_fill_manual(values=c(b_color,b_color,b_color,g_color,l_color,b_color))+
  theme_bw()+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        axis.text=element_blank(),axis.ticks=element_blank(),
        legend.position = "none",
        plot.margin = unit(c(0, 0, 0, 0),"points"),
        panel.margin = unit(c(0, 0, 0, 0), "null")
        )+
  ylim(16.43,18.16) +
  xlim(-88.2,-87.4)+
  geom_text(aes(x=-87.68,y=18.1),label="Barrier")+
  geom_text(aes(x=-87.7,y=17.7),label="Turneffe")+
  geom_text(aes(x=-87.8,y=16.65),label="Glover's")+
  geom_text(aes(x=-87.65,y=17.07),label="Lighthouse")+
  xlab(NULL)+ylab(NULL)+
  annotation_scale(location='br', width_hint=0.3)

#### and with no colors
ggplot(data=barrier)+
  geom_sf(color='black',fill='black')+
  geom_sf(data=turneffe,color='black',fill='black')+
  geom_sf(data=glovers,color='black',fill='black')+
  geom_sf(data=lighthouse,color='black',fill='black')+
  geom_sf(data=sgs_points_sf,size=3, shape=21, fill='white')+
  theme_bw()+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        axis.text=element_blank(),axis.ticks=element_blank(),
        legend.position = "none",
        plot.margin = unit(c(0, 0, 0, 0),"points"),
        panel.margin = unit(c(0, 0, 0, 0), "null")
        )+
  ylim(16.43,18.16) +
  xlim(-88.2,-87.4)+
  geom_text(aes(x=-87.68,y=18.1),label="Barrier")+
  geom_text(aes(x=-87.7,y=17.7),label="Turneffe")+
  geom_text(aes(x=-87.8,y=16.65),label="Glover's")+
  geom_text(aes(x=-87.65,y=17.07),label="Lighthouse")+
  xlab(NULL)+ylab(NULL)+
  annotation_scale(location='br', width_hint=0.3)
```
### Plot with bathymetry:

```{r}
library(marmap)
load('../data/SGS_sites.RData')

#bathy_bze <- getNOAA.bathy(lon1 = -90, lon2 = -86, lat1 = 15, lat2 = 20, resolution = 0.25)
#bathy_xyz <- as.xyz(bathy_bze)

g <- ggplot(barrier)+
  geom_contour_filled(data=bathy_xyz,aes(x=V1,y=V2,z=V3),breaks=c(1000,-1,-10,-20,-30,-40,-50,-60,-5000),alpha=0.5)+
  #scale_fill_manual(values = c("tan","#f7fbff","#deebf7","#c6dbef","#9ecae1","#6baed6","#4292c6","#2171b5","#084594"))+
  #scale_fill_manual(values = c("gray28","#c6dbef","orange","orange2","#9ecae1","#6baed6","#4292c6","#2171b5","#084594"))+
  scale_fill_manual(values = c("gray28","#c6dbef","#9ecae1","#6baed6","#4292c6","#2171b5","#08519c","#08306b"), 
                    name = "Depth(m), \n from NOAA ETOPO 2022")+
  #geom_contour_filled(data=bathy_xyz,aes(x=V1,y=V2,z=V3),breaks=c(-20,-40),color="red",fill="red")+
  #geom_sf(color='black',fill='black')+
  geom_sf(fill="black",aes(color="coral reef"))+
  scale_color_manual(values=c("black"), name = "Coral Cover, \n from Millennium Mapping Project")+
  geom_sf(data=turneffe,color='black',fill='black')+
  geom_sf(data=glovers,color='black',fill='black')+
  geom_sf(data=lighthouse,color='black',fill='black')+
  geom_sf(data=sgs_points_sf,color='black',fill='white',shape=21,size=3)+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank())+
  ylim(16.42,18.16) +
  xlim(-88.2,-87.4)+
  geom_text(aes(x=-87.65,y=18.1),label="Barrier", color="white")+
  geom_text(aes(x=-87.7,y=17.7),label="Turneffe", color="white")+
  geom_text(aes(x=-87.8,y=16.65),label="Glover's", color="white")+
  geom_text(aes(x=-87.65,y=17.05),label="Lighthouse", color="white")+
  xlab(NULL)+ylab(NULL)

g
```

### Use marmap to calculate over-water distances among sites
```{r}
trans1 <- trans.mat(bathy_bze,min.depth=0)
out1 <- lc.dist(trans1,sites_lonlat,res="path")
```



```{r,eval=FALSE}
######## fine scale colors map (for talk)

n_color= '#2b8cbe' #northern (N)
c_color='#a6bddb' #central (C)
t_color='#ece7f2' #turneffe (T)

ggplot(data=barrier)+
  geom_sf(color='black',fill='black')+
  geom_sf(data=turneffe,color='black',fill='black')+
    geom_sf(data=glovers,color='black',fill='black')+
  geom_sf(data=lighthouse,color='black',fill='black')+
  geom_sf(data=filter(sgs_points_sf,
                      Region %in% c("barrier (northern)", "barrier (central)", "Turneffe Atoll")),
          size=3,shape=21, aes(fill=Region))+
  scale_fill_manual(values=c(n_color,c_color,t_color))+
  theme_bw()+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        axis.text=element_blank(),axis.ticks=element_blank(),
        legend.position = "none",
        plot.margin = unit(c(0, 0, 0, 0),"points"),
        panel.margin = unit(c(0, 0, 0, 0), "null")
        )+
  ylim(16.43,18.16) +
  xlim(-88.2,-87.4)+
  geom_text(aes(x=-87.68,y=18.1),label="Barrier")+
  geom_text(aes(x=-87.7,y=17.7),label="Turneffe")+
  geom_text(aes(x=-87.8,y=16.65),label="Glover's")+
  geom_text(aes(x=-87.65,y=17.07),label="Lighthouse")+
  xlab(NULL)+ylab(NULL)+
  annotation_scale(location='br', width_hint=0.3)
```


```{r,eval=FALSE}
######### IBD map (for talk)

ggplot(data=barrier)+
  geom_sf(color='black',fill='black')+
  geom_sf(data=turneffe,color='black',fill='black')+
  geom_sf(data=glovers,color='black',fill='black')+
  geom_sf(data=lighthouse,color='black',fill='black')+
  geom_sf(data=filter(sgs_points_sf,Region %in% c("barrier (northern)", "barrier (central)")),
          size=3, shape=21, fill='black')+
  theme_bw()+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        axis.text=element_blank(),axis.ticks=element_blank(),
        legend.position = "none",
        plot.margin = unit(c(0, 0, 0, 0),"points"),
        panel.margin = unit(c(0, 0, 0, 0), "null")
        )+
  ylim(16.43,18.16) +
  xlim(-88.2,-87.4)+
  geom_text(aes(x=-87.68,y=18.1),label="Barrier")+
  geom_text(aes(x=-87.7,y=17.7),label="Turneffe")+
  geom_text(aes(x=-87.8,y=16.65),label="Glover's")+
  geom_text(aes(x=-87.65,y=17.07),label="Lighthouse")+
  xlab(NULL)+ylab(NULL)+
  annotation_scale(location='br', width_hint=0.3)

```


```{r,eval=FALSE}
plot_map_colors <- ggplot(barrier)+
  geom_sf(color=b_color,fill=b_color)+
  geom_sf(data=turneffe,color=b_color,fill=b_color)+
  geom_sf(data=glovers,color=g_color,fill=g_color)+
  geom_sf(data=lighthouse,color=l_color,fill=l_color)+
#  geom_sf(data=sgs_points_sf,color='black',aes(fill=Region))+
#  scale_fill_manual(values=c(b_color,b_color,b_color,g_color,l_color))+
  geom_sf(data=sgs_points_sf,color='black',fill=,shape=21,size=2)+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        axis.text=element_blank(),axis.ticks=element_blank(),
        panel.background = element_rect(fill = "#f0f0f0"))+
  ylim(16.42,18.16) +
  xlim(-88.2,-87.4)+
  geom_label(aes(x=-87.82,y=17.75),label="Barrier/ \nTurneffe",fill=b_color,alpha=0.1)+
  geom_label(aes(x=-87.8,y=16.65),label="Glover's",fill=g_color,alpha=0.1)+
  geom_label(aes(x=-87.65,y=17.05),label="Lighthouse",fill=l_color,alpha=0.1)+
  xlab(NULL)+ylab(NULL)

plot_map_bw <- ggplot(barrier)+
  geom_sf(color='black',fill='black')+
  geom_sf(data=turneffe,color='black',fill='black')+
  geom_sf(data=glovers,color='black',fill='black')+
  geom_sf(data=lighthouse,color='black',fill='black')+
  geom_sf(data=sgs_points_sf,color='black',fill='white',shape=21,size=2)+
  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
        axis.text=element_blank(),axis.ticks=element_blank(),
        panel.background = element_rect(fill = "#f0f0f0"))+
  ylim(16.42,18.16) +
  xlim(-88.2,-87.4)+
  geom_text(aes(x=-87.65,y=18.1),label="Barrier")+
  geom_text(aes(x=-87.7,y=17.7),label="Turneffe")+
  geom_text(aes(x=-87.8,y=16.65),label="Glover's")+
  geom_text(aes(x=-87.65,y=17.05),label="Lighthouse")+
  xlab(NULL)+ylab(NULL)


# put the atolls together
all_atolls <- st_multipolygon(x=list(glovers,lighthouse,turneffe)) %>%
  st_sfc(crs = 4326)

# plot it all
ggplot(all_atolls)+
  geom_sf()+
  geom_sf(data=barrier)+
  theme(panel.background = element_rect(fill = "lightblue"))

save(all_atolls,barrier,TNC_sim,file='../data/TNC_sim.RData')
```

### Generate plots
1) png file for use by SLiM
2) figure showing reef and the SGS sampling sites, for reference
```{r}
load('../data/TNC_sim.RData')
load('../data/mmp_crop.RData')
load('../data/SGS_sites.RData')

#--------------------------------------------------------------------
# Plot for SLiM
#--------------------------------------------------------------------

g <- ggplot(barrier)+
  geom_sf(color='black',fill='black')+
  geom_sf(data=all_atolls,color='black',fill='black')+
  labs(x=NULL,y=NULL)+
  coord_sf(xlim=c(min_lon,max_lon),ylim=c(min_lat,max_lat),expand=FALSE)+
  theme(axis.ticks = element_blank(),
        axis.text = element_blank(),
        axis.title = element_blank(),
        axis.line = element_blank(),
        legend.position = "none",
        axis.ticks.length = unit(0, "null"),
        axis.ticks.margin = unit(0, "null"),
        legend.margin = unit(0, "null"),
        panel.background = element_rect(fill = "white"),
        panel.grid.major = element_blank(),
        panel.border=element_blank(),
        plot.margin = unit(c(0, 0, 0, 0),"points"),
        panel.margin = unit(c(0, 0, 0, 0), "null"))

# I'm not sure how big it should be, but keep it proportional
ggsave(filename="../data/map_SLiM.png",plot=g,width=round(dist_x)/10,height=round(dist_y)/10,units='cm')


#--------------------------------------------------------------------
#Reference plot
#--------------------------------------------------------------------

h <- ggplot(barrier)+
  geom_sf(color='black',fill='black')+
  geom_sf(data=all_atolls,color='black',fill='black')+
  geom_sf(data=sgs_points_sf,color='blue')+
  coord_sf(xlim=c(min_lon,max_lon),ylim=c(min_lat,max_lat),expand=FALSE)+
  geom_sf_label(data=sgs_points_sf,aes(label=Site),color='blue',hjust=0, vjust=0,size=2)+
  theme(panel.background = element_rect(fill = "lightblue"),
        panel.grid.major = element_blank(),
        panel.border=element_blank(),
        plot.margin = unit(c(0, 0, 0, 0),"points"),
        panel.margin = unit(c(0, 0, 0, 0), "null"))
ggsave(filename="../map_reference.png",plot=h)
```


### Calculate population size
This is Ne for gobies within the previously defined geographic extent.
I'm setting up the model with that many habitable sponges, so assuming habitat was saturated when Naaykens and D'Aloia estimated De.
?? Effective vs actual population size? Can I just use Ne?
```{r}
load('../data/TNC_sim.RData')
load('../data/mmp_crop.RData')
load('../data/SGS_sites.RData')

De_Naaykens <- 749 # effective pop density (individuals/km), Naaykens and D'Aloia 2022. This estimate is assuming random mating.

#--------------------------------------------------------------------
# 1. Get pop density per km^2.
#--------------------------------------------------------------------
# Value from Naaykens and D'Aloia 2022 is per linear km of reef, so first need to convert to per km^2:
# How many km^2 in one linear km of reef?

# Generate the region of that study, i.e., the 41km of barrier reef centered around Carrie Bow.
carriebow_coords <- sites_lonlat[15,] # Carrie Bow is site B16 (15th row of site data)
transect_n <- destPoint(carriebow_coords,0,1000*41/2) # get lat/lon 20.5km N and S of Carrie Bow
transect_s <- destPoint(carriebow_coords,180,1000*41/2)
# create an object that contains all the reef within that lat/lon
parentage_transect <- st_crop(TNC_sim,xmin=-88.08,xmax=-88,ymin=transect_s[2],ymax=transect_n[2])
# here's what it looks like:
ggplot(parentage_transect)+
  geom_sf()+
  geom_sf(data=sgs_points_sf,color='black')+
  geom_sf_text(data=sgs_points_sf,aes(label=Site),color='white',hjust=0, vjust=0,size=2)+
  theme(panel.background = element_rect(fill = "lightblue"),panel.grid.major = element_blank())+
  xlim(c(-88.3,-88))+
  ylim(c(16.5,17))+
  ggtitle('parentage transect')

# Find the area of that region
area_parentage <- st_area(parentage_transect) # calculate the area
area_parentage <- sum(area_parentage) #in m^2
units(area_parentage) <- "km^2" # and convert to km^2

# It's 41km long, so this gives the (average) area of 1km of linear reef, in km^2
area_parentage_1km <- area_parentage/41

#--------------------------------------------------------------------
# 2. Find the total area of reef (km^2) on which sponges will be distributed
#--------------------------------------------------------------------

area_reef <- sum(st_area(barrier))+sum(st_area(all_atolls))
units(area_reef) <- "km^2"
#Looking at a map, ~269km^2 (a box ~16km on each side) seems about right for the total reef area

#--------------------------------------------------------------------
# 3. Total number of individuals on the reef = 419998
#--------------------------------------------------------------------

De_km2 <- De_Naaykens/area_parentage_1km # density per km^2 = (individuals)/(linear km) * (linear km)/(km^2)
Ne_reef <- De_km2*area_reef # total number of individuals on reef
```

## Generate sampling site coordinates for SLiM to use
SLiM's coordinate system is in km, with (0,0) in the SW corner.
```{r}
load('../data/SGS_sites.RData')
load('../data/mmp_crop.RData')

sgs_sites_slim <- data.frame(x = sites_lonlat[,1], y=sites_lonlat[,2])

for(i in 1:nrow(sites_lonlat)){
  sgs_sites_slim$x[i] = distGeo(sites_lonlat[i,],c(min_lon,sites_lonlat[i,2]))/1000
  sgs_sites_slim$y[i] = distGeo(sites_lonlat[i,],c(sites_lonlat[i,1],min_lat))/1000
}

sgs_sites_slim <- mutate(sgs_sites_slim, Site = sgs_points_sf$Site,n_SNP = sgs_points_sf$n_SNP, Region = sgs_points_sf$Region)

write_csv(sgs_sites_slim,file='../data/sgs_sites_SLiM')
```
